Require a chunk size when acknowledging to make sure that the client does not use the outdated cached chunk information before acknowledging it

+++ server/ondisk.go
@@ -10,6 +10,8 @@ import (
 	"regexp"
 	"strconv"
 	"sync"
+
+	"github.com/YuriyNasretdinov/chukcha/protocol"
 )
 
 // TODO: limit the max message size too.
@@ -100,6 +102,11 @@ func (s *OnDisk) getFileDescriptor(chunk string) (*os.File, error) {
 		return nil, fmt.Errorf("create file %q: %s", fp.Name(), err)
 	}
 
+	_, err = fp.Seek(0, io.SeekEnd)
+	if err != nil {
+		return nil, fmt.Errorf("seek file %q until the end: %v", fp.Name(), err)
+	}
+
 	s.fps[chunk] = fp
 	return fp, nil
 }
@@ -150,7 +157,7 @@ func (s *OnDisk) Read(chunk string, off uint64, maxSize uint64, w io.Writer) err
 }
 
 // Ack marks the current chunk as done and deletes it's contents.
-func (s *OnDisk) Ack(chunk string) error {
+func (s *OnDisk) Ack(chunk string, size int64) error {
 	s.Lock()
 	defer s.Unlock()
 
@@ -160,9 +167,13 @@ func (s *OnDisk) Ack(chunk string) error {
 
 	chunkFilename := filepath.Join(s.dirname, chunk)
 
-	_, err := os.Stat(chunkFilename)
+	fi, err := os.Stat(chunkFilename)
 	if err != nil {
 		return fmt.Errorf("stat %q: %w", chunk, err)
+	}
+
+	if fi.Size() > size {
+		return fmt.Errorf("file was not fully processed: the supplied processed size %d is smaller than the chunk file size %d", size, fi.Size())
 	}
 
 	if err := os.Remove(chunkFilename); err != nil {
@@ -178,8 +189,8 @@ func (s *OnDisk) Ack(chunk string) error {
 }
 
 // ListChunks returns the list of current chunks.
-func (s *OnDisk) ListChunks() ([]Chunk, error) {
-	var res []Chunk
+func (s *OnDisk) ListChunks() ([]protocol.Chunk, error) {
+	var res []protocol.Chunk
 
 	dis, err := os.ReadDir(s.dirname)
 	if err != nil {
@@ -187,9 +198,17 @@ func (s *OnDisk) ListChunks() ([]Chunk, error) {
 	}
 
 	for _, di := range dis {
-		c := Chunk{
+		fi, err := di.Info()
+		if errors.Is(err, os.ErrNotExist) {
+			continue
+		} else if err != nil {
+			return nil, fmt.Errorf("reading directory: %v", err)
+		}
+
+		c := protocol.Chunk{
 			Name:     di.Name(),
 			Complete: (di.Name() != s.lastChunk),
+			Size:     uint64(fi.Size()),
 		}
 		res = append(res, c)
 	}
